improve cache hits for multi-project builds
this automatically moves project dependent jar files to most frequently changed layer

e.g. if you have a gradle module using this plugin that depends on another module,
that dependent jar file is now put in the same layer as the classes of the gradle module
where this plugin is used.